REFACTOR: Simplify synchronise profile code

+++ profiles/admin.py
@@ -8,31 +8,18 @@ admin.site.register(Notification)
 
 def synchronize_profile_ids(modeladmin, request, queryset):
     for user in queryset:
-        try:
-            profile = user.profile
-            if profile.id != user.id:
-                try:
-                    conflicting_profile = Profile.objects.get(id=user.id)
-                    conflicting_profile.delete()
-                except Profile.DoesNotExist:
-                    pass
-                profile.id = user.id
-                profile.save()
-        except Profile.DoesNotExist:
-            Profile.objects.create(user=user, id=user.id)
-        except Profile.MultipleObjectsReturned:
-            profiles = Profile.objects.filter(user=user)
+        profiles = Profile.objects.filter(user=user)
+
+        if profiles.exists():
             main_profile = profiles[0]
             for profile in profiles[1:]:
                 profile.delete()
             if main_profile.id != user.id:
-                try:
-                    conflicting_profile = Profile.objects.get(id=user.id)
-                    conflicting_profile.delete()
-                except Profile.DoesNotExist:
-                    pass
+                Profile.objects.filter(id=user.id).delete()
                 main_profile.id = user.id
                 main_profile.save()
+        else:
+            Profile.objects.create(user=user, id=user.id)
 
 
 synchronize_profile_ids.short_description = "Synchronize Profile IDs with User IDs"
